requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
aiomysql>=0.2.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
loguru>=0.7.0
httpx>=0.25.0
pyyaml>=6.0.0
greenlet>=3.3.0
openjiuwen==0.1.7
jinja2>=3.0.0
psycopg2>=2.9.0